// src/app/generated_services/model/updateLessonScheduleDTO.ts
/**
 * CarlsonGracie Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';


export interface UpdateLessonScheduleDTO { 
    title: string;
    description?: string;
    dayOfWeek: DayOfWeek;
    startTime: string;
    duration: string;
    isActive?: boolean;
}
export namespace UpdateLessonScheduleDTO {
}